@@ -8,26 +8,29 @@
     "execa": "^4.0.2",
     "find-yarn-workspace-root": "^2.0.0",
     "fs-extra": "^9.0.1",
+    "tempy": "^0.5.0",
     "yargs-parser": "^18.1.3"
   },
+  "bin": "./bin/yarn-workspace-focus-install.js",
   "devDependencies": {
-    "@babel/core": "^7.10.1",
-    "@babel/preset-env": "^7.10.1",
+    "@babel/core": "^7.10.2",
+    "@babel/preset-env": "^7.10.2",
     "@babel/preset-typescript": "^7.10.1",
     "@kosko/env": "^0.5.2",
     "@socialgouv/eslint-config-recommended": "^1.0.0-beta.4",
     "@socialgouv/eslint-config-typescript": "^1.0.0-beta.4",
     "@socialgouv/yarn-workspace-focus-install": "link:./src",
-    "@types/jest": "^25.2.3",
-    "@types/node": "^14.0.1",
+    "@types/fs-extra": "^9.0.1",
+    "@types/jest": "^26.0.0",
+    "@types/node": "^14.0.13",
     "babel-plugin-module-resolver": "^4.0.0",
-    "eslint": "^7.1.0",
+    "eslint": "^7.2.0",
     "eslint-import-resolver-typescript": "^2.0.0",
     "husky": "^4.2.5",
     "jest": "^26.0.1",
-    "lint-staged": "^10.2.6",
+    "lint-staged": "^10.2.10",
     "prettier": "^2.0.5",
-    "typescript": "^3.9.3"
+    "typescript": "^3.9.5"
   },
   "files": [
     "bin",
